refactor: rename payment_schemas.py and star_service.py
- Rename payment_schemas.py to payment.py
- Rename star_service.py to star_price.py
