+++ core/ioc.py
@@ -2,7 +2,7 @@ from dishka import Provider, Scope, provide
 from core.repositories.user_repo import UserRepository
 from core.repositories.trans_repo import TransactionRepository
 from core.repositories.settings_repo import SettingsRepository
-from core.services.star_service import StarService
+from core.services.star_price import StarService
 from core.services.payment import PaymentService
